Railings + Portal
Portal is still a WIP, I will probably redo it. Did some grunt work on the railings with coliders since we can't auto generate because of the design as it'll give weird results. So instead they all use a box shape like teh original railings. With that, I have succesfully remade every 3D model we need to build a basic game.